+++ include/common.h
@@ -1,0 +1,10 @@
+#ifndef INCLUDE_COMMON_H_
+#define INCLUDE_COMMON_H_
+
+bool isContains(const char *str, char item);
+bool isDigit(char ch);
+bool isSpace(char ch);
+bool isLower(char ch);
+bool isUpper(char ch);
+
+#endif  // INCLUDE_COMMON_H_
